Made a common component for RemoveBaking/delegation + Addressed feedback for related text
