+++ packages/browser-wallet/src/popup/shared/i18n/da.ts
@@ -160,8 +160,8 @@ const t: typeof en = {
             incorrectUrlField: 'Url field was present but did not contain an url',
         },
     },
-    continue: 'Fortsæt',
     cancel: 'Tilbage',
+    continue: 'Fortsæt',
 };
 
 export default t;
